`SqliteDosStorage`: Change default for filepath to be in config folder
The point of the `core.sqlite_dos` storage plugin is to provide a
permanent storage that is easy to setup. However, the default value for
the `filepath` was created in the systems temporary directory, causing
the storage being permanent to not be respected as it was likely to get
cleaned unexpectedly.

Instead, the default is now placed inside the `repository` directory of
the config folder. This is also the location of the default for the
`verdi quicksetup` and `verdi setup` commands. The one difference is that
for the former, the actual folder is based on the chosen profile name,
however, this approach is not possible as the profile name is not
available when the default is created using the `Configuration` model.
Instead, the name takes a random UUID prefixed with `sqlite_dos_` to
indicate what the storage plugin is.