`CalcJob`: Fix MPI behavior if `withmpi` option default is True (#6218)
The documentation states that if neither:

* the `Code` input (through the `withmpi` attribute)
* the `CalcJob` implementation (through `CodeInfo.withpi')
* nor the `metadata.options.withmpi` input

explicitly defines a boolean value, then the implementation falls back
on the _default_ of the `metadata.options.withmpi` input port. However,
the code was actually ignoring this and always falling back to `False`.
This is now corrected, with `False` still as ultimate fallback value in
case a default for `metadata.options.withmpi` is not defined at all.