CLI: Add `--set-as-default` option to `verdi profile setup` (#6221)
In most cases when a new profile is created, a user will expect it to be
the new default. Currently this was not the case for `verdi profile setup`
even if it was the first profile to be created.

The `--set-as-default/--no-set-as-default` flag is added which when
enabled will set the new profile as the default.