`DynamicEntryPointCommandGroup`: Allow entry points to be excluded (#6223)
The `DynamicEntryPointCommandGroup` now checks plugin classes for the
`cli_exposed` class attribute, and when set to `False`, no dynamic
command is generated for the entry point, even if it matches the entry
point filter regex.

This is used for the `core.sqlite_temp` storage plugin, since this is a
storage that is automatically cleaned up after it is garbage collected,
for example when the Python interpreter shuts down. It therefore does
not make sense to allow a profile to be created using this storage
plugin through `verdi profile setup`.