`verdi profile delete`: Make it storage plugin agnostic
The `verdi profile delete` implementation was hardcoded to work only for
the `core.psql_dos` storage plugin. However, it is possible for profiles
to use different storage plugins, causing the command to except.

The command now forwards to `Config.delete_profile`. This method gets
the storage plugin defined for the profile and constructs an instance of
it for the given profile. It then calls `delete` on the storage instance
if the `delete_storage` argument is set to `True`.

The new `delete` method is defined on the `StorageBackend` abstract base
class. It is intentionally not made abstract, but rather explicitly
raises `NotImplementedError`. Otherwise existing storage plugins outside
of `aiida-core` would no longer be instantiable before they implement
the method.

The various options of `verdi profile delete` to control what parts of
the storage are deleted, are simplified to just a single switch that
controls whether the storage is deleted or not. The switch is not
defined at all by default. The user will have to explicitly select
either `--delete-data` or `--keep-data`, or they will be prompted what
value to choose. If the `--force` flag is specified and no explicit flag
to delete or keep the data, the command raises.

The command now also properly detects if the delete profile was the
default, in which case another profile is marked as the new default if
any profiles remain at all.