Docs: Fix instructions for inspecting archive files
The instructions were incorrect and caused exceptions:

* The `profile_context ` would except if another profile would already be
  loaded, unless `allow_switch=True` is passed explicitly.
* The profile will use a relative path for the archive in the profile
  config. This will cause an exception when the path cannot be found.
* No default user would be set for the profile, causing an exception
  when a new ORM entity would be instantiated.

The instructions are fixed by suggesting to use the new CLI command
`verdi profile setup` to create a profile using the `core.sqlite_zip`
storage. This allows the profile to be used like any other profile,
simplifying the instructions, and fixing the problems above.