Docs: Remove `aiida.manage.configuration.load_documentation_profile` (#6231)
This function was used by the ReadTheDocs build in `docs/source/conf.py`
and the `utils/validate_consistency.py` function to automatically build
the `verdi` documentation. This was necessary when Django was still used
as a storage backend as merely importing the ORM modules would try to
connect to the database.

Since v2.0, Django is no longer supported and only Sqlalchemy is used
which no longer establishes database connection upon module imports.
Therefore the function `load_documentation_profile` which only existed
to create a dummy profile with database connection is no longer needed
and is hereby removed.

The `docs/source/conf.py` still needs to manually import the various
modules of the `aiida` package for the auto API doc to work.