Remove post release action for uploading pot to transifex
The action was for building and pushing the pot files to transifex for
translation.
The repo that manage the public of translated doc is https://github.com/unkcpz/aiida-core-i18n.
The pull strategy is implemented there to keep the pot files synchronized, which should be
more robust and decouple from rely on the release action push of pot files.