CLI: Fix `verdi process repair` not actually repairing
The command would use `echo_critical` to report an inconsistency when
found, however, this function immediately exists the interpreter and so
the code that would perform the fix is unreachable. The command now
simply emits a warning if an inconsistency is detected, except in the
case of a `--dry-run` when the command still immediately exits with a
non-zero exit code.