Dependencies: Fix incompatibility with `spglib>=2.3`
The `spglib` library deprecated support for `ase.Atoms` objects a long
time ago and it was finally removed in the v2.3 release. However, the
`aiida.orm.nodes.data.structure.ase_refine_cell` function was still
using the deprecated functionality. The `Atoms` objects are now replaced
with the plain tuples that `spglib` expects.