CLI: Speed up tab-completion by lazily importing `Config` (#6275)
The `Config` class was recently changed to use `pydantic` which adds
a non-negligible load time to `verdi`, which affects the responsiveness
of tab-completion. To restore this, the `Config` class cannot be imported
upon loading `verdi` and so its exposure in the parent module had to be
removed. Instead of importing from `aiida.manage.configuration` it now
has to be imported from the full path `aiida.manage.configuration.config`.