ORM: Fix `ProcessNode.get_builder_restart`
The `get_builder_restart` would fail if the process input specification
contained the inputs of a `CalcJob`. The method would first assign the
regular inputs to the builder followed by the metadata. However, each
time the builder is updated, the validation is triggered. The normal
inputs do not contain the required metadata, such as the `resources`
option, causing the validation to fail.

The solution is to update the builder with the complete set of inputs in
one go, merging the normal inputs with the metadata inputs prior to that.