CLI: Let `-v/--verbosity` only affect `aiida` and `verdi` loggers (#6286)
In 42c5ab101c77a7e278bddb87d0055b907db42dc3, the logging of the CLI was
reworked, specifically the functionality of the `-v/--verbosity`. The
decision was made to have the log level, specified through this option,
be applied to all loggers. While this makes sense, this had some
unintended consequences.

The `verdi` commands and the Python API themselves often use `INFO` log
messages to provide additional useful information. But when enabled
using `-v INFO`, they would be drowned out by info messages of other
loggers, most notably by the `sqlalchemy` logger. Since we cannot
control the level at which these messages of third-party packages are
emitted, we are forced to not update the loglevel for these loggers.

The downside is that now the loggers are no longer all controlled at
once through a single option. But it is doubtful that anyone was relying
on this anyway, since the most relevant information comes from the
`aiida` logger. And it is still possible to change the log level of
external loggers through the configuration options.