ORM: Undo deprecation of `Code.get_description`
The `get_description` method was unjustly deprecated in favor of the
`description` property. However, the latter is just a direct reference
to the `description` column of the database model. This is an optional
string that can be defined by the user when the instance is created.
 The `get_description` method rather, which is defined on the `Node`
base class, is intended to return a description composed from relevant
properties of the node. These are therefore not the same.

The deprecation is undone and it now returns `Code.full_label`.