Logging: Add the `disk_objectstore` logger to the config
The `disk_objectstore` logger is added to the default logging
configuration in `aiida.common.log.get_logging_config`. The log level
can be controlled through the `logging.disk_objectstore_loglevel` config
option. Its default is intentionally set to `INFO` because information
during backup is logged on that level and we want it to be shown by
default.