Refactor: Abstract message broker functionality
RabbitMQ is used as a message broker to allow processes to be
distributed across multiple daemon workers. Setting up RabbitMQ as a
service can be a hurdle for users and for simple use cases the
additional scaling it provides is not always necessary.

The idea is to provide a light version of AiiDA that doesn't require a
message broker, making it easy to setup and install to allow users to
quickly familiarize themselves with AiiDA before committing to a more
production ready setup.

Currently, the RabbitMQ functionality is a bit dispersed throughout the
codebase, making it difficult to make its use optional. Therefore, the
first step is to refactor the code and centralize all RabbitMQ specific
functionality.

The `aiida.brokers` module is introduced, which provides the abstract
class `Broker`. This is the interface that a message broker, such as
RabbitMQ should implement. The existing RabbitMQ implementation is added
to the `RabbitmqBroker` class.

The `Manager` class now provides the `get_broker` method which will load
the broker instance based on the plugin that is defined by the profile.
This is an entry point that should be registered in the `aiida.brokers`
group.

For now, a `NocommsBroker` is implemented which should serve the case
when no message broker is available or needed. It is not quite sure if
the alternative of just having `Manager.get_broker` returning `None`
would be simpler and ultimately better.

These changes are just a first refactoring that don't change any
functionality and everything is backwards compatible. In the future,
functionality can be added to allow creating profiles that do not define
a broker. These can then be used to run processes locally in the same
Python interpreter in which they are launched.