CLI: Add the `verdi blitz` command
This command will provide a one-stop-shop to create a new profile that
requires no services (no PostgreSQL nor RabbitMQ) and is as much
ready-to-go as is possible.

The profile created uses the `core.sqlite_dos` plugin for the storage
and sets `core.no_comms` as the broker. The profile is automatically set
to be the default and a number of options are configured to sensible
defaults. Finally the localhost is configured as a `Computer`.

This command will make it a lot easier for users to try AiiDA as
services are typically difficult to install and are not really necessary
for a sandbox so potential users can play around to get a sense of how
things work. With this addition, going from scratch to running a calcjob
takes three commands in total:

    pip install aiida-core
    verdi blitz
    verdi devel launch-add

Of course, if a user wants to run other codes, they may have to install
another plugin and install their pre-requisites.