Docs: Reorganize the tutorial content
The tutorials are moved to a separate page. A link is added to the
`aiida-tutorials` repository as a source of additional tutorials. In the
future, valuable tutorial information will be taken from there and
directly integrated in the new tutorials page of the `aiida-core` docs.

The "What is AiiDA" section is moved to the index page of the
introduction.