CLI: Fix logging not showing in `verdi daemon worker` (#6301)
The `verdi daemon worker` will launch a single daemon worker in the
foreground which is useful for debugging. However, in this case, the
logging was not configured to write to stdout which renders the command
essentially moot.

The `aiida.engine.daemon.worker.start_worker` function now takes the
argument `foreground` which is `False` by default, but when set to
`True`, the logging is configured to write to stdout instead of the
daemon log file.