👌Improve: Spruce up `orm.Entity.fields` interface
This commit first ensures that all query operations are supported in the
new syntax, including negation. The commit also separates the operations
by field type via subclassing. It introduces a uniform `orm.fields.add_field`
API as an abstraction of the type-dependent allocation of
`orm.fields.QbField` flavors.

A distinction is made in the API between database columns and attribute
fields via the `is_attribute` parameter. If set to `True` (the default), the field
is considered an attribute field and as such, `QbField().backend_key` will
return `'attribute.key'` in accordance with the `QueryBuilder` mechanics.
Note that _only_ in the case of attribute fields, `backend_key` may be
aliased via the `alias` parameter. This is done to avoid restrictions on
database field naming while maintaining a pythonic interface in the
frontend via the `key` parameter, which is validated as pythonic
(supporting dot notation).