Engine: Fix paused work chains not showing it in process status (#6206)
When child processes of a `WorkChain` are completed, the process status
of the work chain is updated. However, the `_update_process_status` was
not taking into account whether the work chain was paused and would
simply override the paused message. This would result in the process
status being shown as `Waiting`, but actually the process is paused and
so it would look stuck.

Instead, the method now simply updates the `_pre_paused_status` with the
new status in case the work chain is paused. This way, it will continue
to be shown as paused, but as soon as it is resumed, the new status will
be correctly set.