Docker: Use the current checkout for `aiida-core-dev` (#6314)
The `aiida-core-dev` image was always using the default branch off the
main repository to build against. This would result in PRs failing if
they included changes that affect the tests because the tests would not
be run against the code of the branch.

Instead, the current checked out code is now copied to the container.
For situations where the volume is not persistent, as for example when
run with k8s, the main repo default branch is cloned as a fallback.