Config: Check current working directory for existing config dir
Currently, the location of the configuration directory could only be
controlled through the `AIIDA_PATH` environment variable. Here, the
heuristic is updated to also consider the current working directory or
any of its parent directories in case `AIIDA_PATH` is not set. This
provides another easy way for users to maintain multiple separate AiiDA
instances.