Caching: `NodeCaching._get_objects_to_hash` return type to `dict`
When debugging the caching functionality, it is often useful to look at
the objects that are used to compute the hash, which is returned by the
`NodeCaching._get_objects_to_hash`. It returning a list, makes it
difficult to identify what each object represents. By changing the
return type to a dictionary, the key for each object allows to give a
useful hint as to what it represents, making debugging easier.