CLI: Always do hard reset in `verdi daemon restart` (#6317)
By default, `verdi daemon restart` would perform a "soft" restart of the
daemon, meaning that just the daemon workers would be restarted and not
the daemonizer process itself. The original reasoning was that a full
reset was usually not necessary and would take more time than a soft
restart.

However, exactly when a full restart is necessary is not easy to define
and so by default users would be recommended to run with the `--reset`
flag to perform a full restart. In practice, this would add complexity
for little gain because the second extra it takes to perform a full
restart is negligible in most situations.

Therefore, the `--reset` flag, as well as the `--no-wait` and `--timeout`
flags which only make sense in the case of a soft reset, are deprecated
and the command now always performs a full restart.