Caching: Rename `get_hash` to `compute_hash`
The `get_hash` method was actually recomputing the hash and not just
returning the hash stored in the extras that was computed when the node
got stored. The `get_hash` and `_get_hash` methods are renamed to
`compute_hash` and `_compute_hash`, respectively. This allows `get_hash`
to be reimplemented to simply return the hash stored in the extras.

Although technically speaking this is changing the implementation of
public facing API, if users even use `get_hash()` it will have been to
retrieve the hash for an already stored node, which is therefore equal
to just returning the pre-computed hash stored in the extras. So
effectively this should not be a breaking change.